--- game/src/main/kotlin/content/area/kandarin/ardougne/west_ardougne/Man.kts
@@ -1,17 +1,0 @@
-package content.area.kandarin.ardougne.west_ardougne
-
-import content.entity.player.dialogue.*
-import content.entity.player.dialogue.type.npc
-import content.entity.player.dialogue.type.player
-import world.gregs.voidps.engine.entity.character.npc.npcOperate
-
-npcOperate("Talk-to","w_ardougnecitizen3") {
-    player<Happy>("Good day.")
-    npc<Surprised>("An outsider! Can you get me out of this hell hole?")
-    player<Sad>("Sorry, that's not what I'm here to do.")
-}
-
-npcOperate("Talk-to","w_ardougnecitizen4") {
-    player<Happy>("Hello there.")
-    npc<Angry>("Go away. People from the outside shut us in like animals. I have nothing to say to you.")
-}
